Moving passivation to an util class (changes ejb3.stateful.passivation)
